pyproject.toml
[project]
name = "stock-ml-predictor"
version = "1.0.0"
description = "Machine learning project for stock price prediction using yfinance data"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "yfinance>=0.2.28",
    "pandas>=1.5.0",
    "numpy>=1.23.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "joblib>=1.3.0",
    "pyyaml>=6.0",
    "ta>=0.10.2",
    "plotly>=5.17.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"